requirements.txt
asgiref==3.9.1
channels==4.0.0
channels-redis==4.0.0
attrs==25.3.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.6
cryptography==46.0.5
dj-rest-auth==7.2.0
Django==5.2.5
django-allauth==65.15.0
django-cors-headers==4.7.0
django-debug-toolbar==6.0.0
django-filter==25.2
django-lifecycle==1.2.4
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.28.0
drf-spectacular-sidecar==2025.8.1
gunicorn==23.0.0
idna==3.11
inflection==0.5.1
jsonschema==4.25.1
jsonschema-specifications==2025.4.1
oauthlib==3.3.1
packaging==25.0
pycparser==3.0
PyJWT==2.10.1
PyMySQL==1.1.2
python-dotenv==1.1.1
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.5
rpds-py==0.27.0
sqlparse==0.5.3
typing_extensions==4.14.1
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.6.3
whitenoise==6.9.0
